+++ CHANGELOG.md
@@ -10,6 +10,10 @@
 [Will McGinty](https://github.com/wmcginty)
 [#30](https://github.com/BottleRocketStudios/iOS-Scotty/pull/30)
 
+* Added Swift Package Manager support.
+[Brian Miller](https://gitthub.com/jobsismyhomeboy)
+[#28](https://github.com/BottleRocketStudios/iOS-Scotty/issues/28)
+
 ##### Bug Fixes
 
 * None.